requirements.txt
fastapi
uvicorn
openai>=1.3.5
python-multipart
python-dotenv
tiktoken
Pillow
matplotlib